# api/openapi/ogen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package ogen

import (
	"time"
)

// CreateDailyMealBadRequest is response for CreateDailyMeal operation.
type CreateDailyMealBadRequest struct{}

func (*CreateDailyMealBadRequest) createDailyMealRes() {}

// Ref: #/components/schemas/CreateDailyMealRequest
type CreateDailyMealRequest struct {
	Name     string `json:"name"`
	Calories int    `json:"calories"`
}

// GetName returns the value of Name.
func (s *CreateDailyMealRequest) GetName() string {
	return s.Name
}

// GetCalories returns the value of Calories.
func (s *CreateDailyMealRequest) GetCalories() int {
	return s.Calories
}

// SetName sets the value of Name.
func (s *CreateDailyMealRequest) SetName(val string) {
	s.Name = val
}

// SetCalories sets the value of Calories.
func (s *CreateDailyMealRequest) SetCalories(val int) {
	s.Calories = val
}

// Ref: #/components/schemas/Error
type Error struct {
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *Error) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *Error) SetMessage(val string) {
	s.Message = val
}

func (*Error) getDailyMealRes() {}

// ListDailyMealsInternalServerError is response for ListDailyMeals operation.
type ListDailyMealsInternalServerError struct{}

func (*ListDailyMealsInternalServerError) listDailyMealsRes() {}

type ListDailyMealsOKApplicationJSON []Meal

func (*ListDailyMealsOKApplicationJSON) listDailyMealsRes() {}

// Ref: #/components/schemas/Meal
type Meal struct {
	ID        int64     `json:"id"`
	Name      string    `json:"name"`
	Calories  int       `json:"calories"`
	CreatedAt time.Time `json:"created_at"`
}

// GetID returns the value of ID.
func (s *Meal) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *Meal) GetName() string {
	return s.Name
}

// GetCalories returns the value of Calories.
func (s *Meal) GetCalories() int {
	return s.Calories
}

// GetCreatedAt returns the value of CreatedAt.
func (s *Meal) GetCreatedAt() time.Time {
	return s.CreatedAt
}

// SetID sets the value of ID.
func (s *Meal) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Meal) SetName(val string) {
	s.Name = val
}

// SetCalories sets the value of Calories.
func (s *Meal) SetCalories(val int) {
	s.Calories = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *Meal) SetCreatedAt(val time.Time) {
	s.CreatedAt = val
}

func (*Meal) createDailyMealRes() {}
func (*Meal) getDailyMealRes()    {}